+++ test/typewriter_controller_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:typewritertext/typewritertext.dart';
+import 'package:typewritertext/v3/typewriter.dart';
 
 void main() {
   group('TypeWriterController', () {
@@ -15,6 +15,7 @@ void main() {
     tearDown(() {
       controller.dispose();
     });
+
     test('value', () async {
       controller.addListener(() {
         expect(
@@ -37,6 +38,13 @@ void main() {
       });
 
       await controller.start();
+
+      expect(controller.value.index == 'Hello World'.length - 1, isTrue);
+      expect(controller.value.text, equals('World'));
+
+      await controller.resume();
+      expect(controller.value.index == 'Hello World'.length - 1, isTrue);
+      expect(controller.value.text, equals('World'));
     });
   });
 }