+++ test/typewriter_value_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:typewritertext/typewritertext.dart';
+import 'package:typewritertext/v3/typewriter.dart';
 
 void main() {
   group('TypeWriterValue', () {
@@ -14,6 +14,10 @@ void main() {
       expect(value.text, equals('slasher'));
       value.index = 24;
       expect(value.text, equals('the'));
+      expect(
+          value.toString(),
+          equals(
+              'TypeWriterValue{data: [the, hash, slinging, slasher], text: the, index: 2}'));
     });
     test('length', () {
       expect(value.length, equals(22));